Fixa dina fucking varningar

+++ DV1572/DV1572/ourEngine/core/Picking.cpp
@@ -232,7 +232,7 @@ Shape * Picking::getPicked(Camera * c, ID3D11RenderTargetView*&RTV, ID3D11DepthS
 	
 	for (size_t i = 0; i < DX::g_instanceGroupsPicking.size() && indexInPickingQueue == 0; i++)
 	{
-		long nrOfShapes = DX::g_instanceGroupsPicking[i].index.size();
+		long nrOfShapes = static_cast<long>(DX::g_instanceGroupsPicking[i].index.size());
 		if (index < nrOfShapes)
 		{
 			indexInPickingQueue = DX::g_instanceGroupsPicking[i].index[index];
